+++ previous-versions/sync-for-commerce-version-1/pkg/models/operations/getcommercepayment.go
@@ -8,7 +8,9 @@ import (
 )
 
 type GetCommercePaymentRequest struct {
-	CompanyID    string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a company.
+	CompanyID string `pathParam:"style=simple,explode=false,name=companyId"`
+	// Unique identifier for a connection.
 	ConnectionID string `pathParam:"style=simple,explode=false,name=connectionId"`
 	// Unique identifier for a payment.
 	PaymentID string `pathParam:"style=simple,explode=false,name=paymentId"`
@@ -38,11 +40,14 @@ func (o *GetCommercePaymentRequest) GetPaymentID() string {
 type GetCommercePaymentResponse struct {
 	// OK
 	CommercePayment *shared.CommercePayment
-	ContentType     string
+	// HTTP response content type for this operation
+	ContentType string
 	// Your API request was not properly authorized.
 	ErrorMessage *shared.ErrorMessage
-	StatusCode   int
-	RawResponse  *http.Response
+	// HTTP response status code for this operation
+	StatusCode int
+	// Raw HTTP response; suitable for custom response parsing
+	RawResponse *http.Response
 }
 
 func (o *GetCommercePaymentResponse) GetCommercePayment() *shared.CommercePayment {