+++ previous-versions/sync-for-commerce-version-1/pkg/models/shared/accountingaccount.go
@@ -3,7 +3,8 @@
 package shared
 
 import (
-	"github.com/codatio/client-sdk-go/previous-versions/sync-for-commerce-version-1/pkg/types"
+	"github.com/codatio/client-sdk-go/previous-versions/sync-for-commerce-version-1/pkg/utils"
+	"github.com/ericlagergren/decimal"
 )
 
 type AccountingAccountMetadata struct {
@@ -117,7 +118,7 @@ type AccountingAccount struct {
 	// There are only a very small number of edge cases where this currency code is returned by the Codat system.
 	Currency *string `json:"currency,omitempty"`
 	// Current balance in the account.
-	CurrentBalance *types.Decimal `json:"currentBalance,omitempty"`
+	CurrentBalance *decimal.Big `decimal:"number" json:"currentBalance,omitempty"`
 	// Description for the account.
 	Description *string `json:"description,omitempty"`
 	// Full category of the account.
@@ -142,10 +143,25 @@ type AccountingAccount struct {
 	SourceModifiedDate *string `json:"sourceModifiedDate,omitempty"`
 	// Status of the account
 	Status *AccountStatus `json:"status,omitempty"`
+	// Supplemental data is additional data you can include in our standard data types.
+	//
+	// It is referenced as a configured dynamic key value pair that is unique to the accounting platform. [Learn more](https://docs.codat.io/using-the-api/supplemental-data/overview) about supplemental data.
+	SupplementalData *SupplementalData `json:"supplementalData,omitempty"`
 	// Type of account
 	Type *AccountType `json:"type,omitempty"`
 	// The validDatatypeLinks can be used to determine whether an account can be correctly mapped to another object; for example, accounts with a `type` of `income` might only support being used on an Invoice and Direct Income. For more information, see [Valid Data Type Links](/accounting-api#/schemas/ValidDataTypeLinks).
 	ValidDatatypeLinks []AccountingAccountValidDataTypeLinks `json:"validDatatypeLinks,omitempty"`
+}
+
+func (a AccountingAccount) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(a, "", false)
+}
+
+func (a *AccountingAccount) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &a, "", false, false); err != nil {
+		return err
+	}
+	return nil
 }
 
 func (o *AccountingAccount) GetCurrency() *string {
@@ -155,7 +171,7 @@ func (o *AccountingAccount) GetCurrency() *string {
 	return o.Currency
 }
 
-func (o *AccountingAccount) GetCurrentBalance() *types.Decimal {
+func (o *AccountingAccount) GetCurrentBalance() *decimal.Big {
 	if o == nil {
 		return nil
 	}
@@ -237,6 +253,13 @@ func (o *AccountingAccount) GetStatus() *AccountStatus {
 		return nil
 	}
 	return o.Status
+}
+
+func (o *AccountingAccount) GetSupplementalData() *SupplementalData {
+	if o == nil {
+		return nil
+	}
+	return o.SupplementalData
 }
 
 func (o *AccountingAccount) GetType() *AccountType {