+++ previous-versions/sync-for-commerce-version-1/pkg/models/shared/invoicelineitem.go
@@ -3,7 +3,7 @@
 package shared
 
 import (
-	"github.com/codatio/client-sdk-go/previous-versions/sync-for-commerce-version-1/pkg/types"
+	"github.com/codatio/client-sdk-go/previous-versions/sync-for-commerce-version-1/pkg/utils"
 	"github.com/ericlagergren/decimal"
 )
 
@@ -13,22 +13,23 @@ type InvoiceLineItem struct {
 	// Friendly name of the goods or services provided.
 	Description *string `json:"description,omitempty"`
 	// Numerical value of any discounts applied.
-	DiscountAmount *types.Decimal `json:"discountAmount,omitempty"`
+	DiscountAmount *decimal.Big `decimal:"number" json:"discountAmount,omitempty"`
 	// Percentage rate (from 0 to 100) of any discounts applied to the unit amount.
-	DiscountPercentage *types.Decimal `json:"discountPercentage,omitempty"`
-	IsDirectIncome     *bool          `json:"isDirectIncome,omitempty"`
+	DiscountPercentage *decimal.Big `decimal:"number" json:"discountPercentage,omitempty"`
+	// The invoice is a direct income if `True`.
+	IsDirectIncome *bool `json:"isDirectIncome,omitempty"`
 	// Reference to the product, service type, or inventory item to which the direct cost is linked.
 	ItemRef *ItemRef `json:"itemRef,omitempty"`
 	// Number of units of goods or services provided.
-	Quantity types.Decimal `json:"quantity"`
+	Quantity *decimal.Big `decimal:"number" json:"quantity"`
 	// Amount of the line, inclusive of discounts but exclusive of tax.
-	SubTotal *types.Decimal `json:"subTotal,omitempty"`
+	SubTotal *decimal.Big `decimal:"number" json:"subTotal,omitempty"`
 	// Amount of tax for the line.
-	TaxAmount *types.Decimal `json:"taxAmount,omitempty"`
+	TaxAmount *decimal.Big `decimal:"number" json:"taxAmount,omitempty"`
 	// Reference to the tax rate to which the line item is linked.
 	TaxRateRef *TaxRateRef `json:"taxRateRef,omitempty"`
 	// Total amount of the line, including tax. When pushing invoices to Xero, the total amount is exclusive of tax to allow automatic calculations if a tax rate or tax amount is not specified.
-	TotalAmount *types.Decimal `json:"totalAmount,omitempty"`
+	TotalAmount *decimal.Big `decimal:"number" json:"totalAmount,omitempty"`
 	// Categories, and a project and customer, against which the item is tracked.
 	Tracking *Tracking `json:"tracking,omitempty"`
 	// Reference to the tracking categories to which the line item is linked.
@@ -36,7 +37,18 @@ type InvoiceLineItem struct {
 	// Deprecated field: This will be removed in a future release, please migrate away from it as soon as possible.
 	TrackingCategoryRefs []TrackingCategoryRefsitems `json:"trackingCategoryRefs,omitempty"`
 	// Price of each unit of goods or services.
-	UnitAmount types.Decimal `json:"unitAmount"`
+	UnitAmount *decimal.Big `decimal:"number" json:"unitAmount"`
+}
+
+func (i InvoiceLineItem) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(i, "", false)
+}
+
+func (i *InvoiceLineItem) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &i, "", false, false); err != nil {
+		return err
+	}
+	return nil
 }
 
 func (o *InvoiceLineItem) GetAccountRef() *AccountRef {
@@ -53,14 +65,14 @@ func (o *InvoiceLineItem) GetDescription() *string {
 	return o.Description
 }
 
-func (o *InvoiceLineItem) GetDiscountAmount() *types.Decimal {
+func (o *InvoiceLineItem) GetDiscountAmount() *decimal.Big {
 	if o == nil {
 		return nil
 	}
 	return o.DiscountAmount
 }
 
-func (o *InvoiceLineItem) GetDiscountPercentage() *types.Decimal {
+func (o *InvoiceLineItem) GetDiscountPercentage() *decimal.Big {
 	if o == nil {
 		return nil
 	}
@@ -81,21 +93,21 @@ func (o *InvoiceLineItem) GetItemRef() *ItemRef {
 	return o.ItemRef
 }
 
-func (o *InvoiceLineItem) GetQuantity() types.Decimal {
+func (o *InvoiceLineItem) GetQuantity() *decimal.Big {
 	if o == nil {
-		return types.Decimal{Big: *(new(decimal.Big).SetFloat64(0.0))}
+		return new(decimal.Big).SetFloat64(0.0)
 	}
 	return o.Quantity
 }
 
-func (o *InvoiceLineItem) GetSubTotal() *types.Decimal {
+func (o *InvoiceLineItem) GetSubTotal() *decimal.Big {
 	if o == nil {
 		return nil
 	}
 	return o.SubTotal
 }
 
-func (o *InvoiceLineItem) GetTaxAmount() *types.Decimal {
+func (o *InvoiceLineItem) GetTaxAmount() *decimal.Big {
 	if o == nil {
 		return nil
 	}
@@ -109,7 +121,7 @@ func (o *InvoiceLineItem) GetTaxRateRef() *TaxRateRef {
 	return o.TaxRateRef
 }
 
-func (o *InvoiceLineItem) GetTotalAmount() *types.Decimal {
+func (o *InvoiceLineItem) GetTotalAmount() *decimal.Big {
 	if o == nil {
 		return nil
 	}
@@ -130,9 +142,9 @@ func (o *InvoiceLineItem) GetTrackingCategoryRefs() []TrackingCategoryRefsitems 
 	return o.TrackingCategoryRefs
 }
 
-func (o *InvoiceLineItem) GetUnitAmount() types.Decimal {
+func (o *InvoiceLineItem) GetUnitAmount() *decimal.Big {
 	if o == nil {
-		return types.Decimal{Big: *(new(decimal.Big).SetFloat64(0.0))}
+		return new(decimal.Big).SetFloat64(0.0)
 	}
 	return o.UnitAmount
 }