+++ previous-versions/sync-for-commerce-version-1/pkg/models/shared/servicecharge.go
@@ -3,7 +3,8 @@
 package shared
 
 import (
-	"github.com/codatio/client-sdk-go/previous-versions/sync-for-commerce-version-1/pkg/types"
+	"github.com/codatio/client-sdk-go/previous-versions/sync-for-commerce-version-1/pkg/utils"
+	"github.com/ericlagergren/decimal"
 )
 
 // ServiceChargeTaxComponentAllocationTaxComponentRef - Taxes rates reference object depending on the rates being available on source commerce package.
@@ -30,12 +31,23 @@ func (o *ServiceChargeTaxComponentAllocationTaxComponentRef) GetName() string {
 
 type ServiceChargeTaxComponentAllocation struct {
 	// Tax amount on order line sale as available from source commerce platform.
-	Rate *types.Decimal `json:"rate,omitempty"`
+	Rate *decimal.Big `decimal:"number" json:"rate,omitempty"`
 	// Taxes rates reference object depending on the rates being available on source commerce package.
 	TaxComponentRef *ServiceChargeTaxComponentAllocationTaxComponentRef `json:"taxComponentRef,omitempty"`
 }
 
-func (o *ServiceChargeTaxComponentAllocation) GetRate() *types.Decimal {
+func (s ServiceChargeTaxComponentAllocation) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(s, "", false)
+}
+
+func (s *ServiceChargeTaxComponentAllocation) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &s, "", false, false); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (o *ServiceChargeTaxComponentAllocation) GetRate() *decimal.Big {
 	if o == nil {
 		return nil
 	}
@@ -55,15 +67,26 @@ type ServiceCharge struct {
 	// The number of times the charge is charged.
 	Quantity *int64 `json:"quantity,omitempty"`
 	// Amount of the service charge that is tax.
-	TaxAmount *types.Decimal `json:"taxAmount,omitempty"`
+	TaxAmount *decimal.Big `decimal:"number" json:"taxAmount,omitempty"`
 	// Percentage rate (from 0 to 100) of any tax applied to the service charge.
-	TaxPercentage *types.Decimal `json:"taxPercentage,omitempty"`
+	TaxPercentage *decimal.Big `decimal:"number" json:"taxPercentage,omitempty"`
 	// Taxes breakdown as applied to service charges.
 	Taxes []ServiceChargeTaxComponentAllocation `json:"taxes,omitempty"`
-	// Total service charge, including taxes.
-	TotalAmount *types.Decimal `json:"totalAmount,omitempty"`
+	// Total amount of the service charge, including tax.
+	TotalAmount *decimal.Big `decimal:"number" json:"totalAmount,omitempty"`
 	// The type of the service charge.
 	Type *ServiceChargeType `json:"type,omitempty"`
+}
+
+func (s ServiceCharge) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(s, "", false)
+}
+
+func (s *ServiceCharge) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &s, "", false, false); err != nil {
+		return err
+	}
+	return nil
 }
 
 func (o *ServiceCharge) GetDescription() *string {
@@ -80,14 +103,14 @@ func (o *ServiceCharge) GetQuantity() *int64 {
 	return o.Quantity
 }
 
-func (o *ServiceCharge) GetTaxAmount() *types.Decimal {
+func (o *ServiceCharge) GetTaxAmount() *decimal.Big {
 	if o == nil {
 		return nil
 	}
 	return o.TaxAmount
 }
 
-func (o *ServiceCharge) GetTaxPercentage() *types.Decimal {
+func (o *ServiceCharge) GetTaxPercentage() *decimal.Big {
 	if o == nil {
 		return nil
 	}
@@ -101,7 +124,7 @@ func (o *ServiceCharge) GetTaxes() []ServiceChargeTaxComponentAllocation {
 	return o.Taxes
 }
 
-func (o *ServiceCharge) GetTotalAmount() *types.Decimal {
+func (o *ServiceCharge) GetTotalAmount() *decimal.Big {
 	if o == nil {
 		return nil
 	}